Translate some dynamic strings into Chinese
Many of these more dynamic strings require an alteration of the
order of the sentence from Spanish and English. While this was
easy with how we created the text for the tooltips on the meters
in the neighborhood details, it was not as easy for elements such
as the "About [number] minutes via bus from [location]" or the
"Take [transit route] in [number] minutes". These will require
a restructuring of those elements to make the grammar more
flexible, as well as clearer translations from the client so
that the developer does not have to guess as to the order and
proper grammar.