Convert profiles created before translation to use i18next phrase
The translation of the 'from' dropdown menu relies on destination
purposes contain the 'TripPurpose.' for use with reacti18next.
Profiles created after the addition of this translation work
with this setup as that addition occurs in EditProfile; however,
profiles created before and stored with purposes that do not contain
the 'TripPurpose.' prefix do not get properly translated. This commit
adds logic to check for such instances and convert any old profiles
when populating the menu's options.