Add new font size to account for wrapping of Chinese text
The initial font size for nieghborhood-summary__trajectory in the
route card was 200. This size was too big and created wrapping when
the user would change to Chinese. Changing to 100 was too small and
difficult to read in English and Spanish. This commit adds a new
font size that is in between the two to use in this instance.