Replace login and edit profile with anonymous-only authentication
This commit replaces the Amplify components from custom-sign-in,
index, edit-profile, and custom-header-bar with an anonymous-only
authentication wrapper. It also removes reference to checks for
isCounselor and isAnonymous, as well as other parts of client
profiles (such as email validation and profile confirmation) that
are not relevant when the client is anonymous.