Add django-sesame and connect to custom-sign-in on frontend
This commit sets up django-sesame and uses it within a view to
generate a magic link unique to the user. If the provided user is
found, it will send an email containing that link; otherwise, it
will fail silently rather than sending a 500 error back in order to
protect the privacy of those using the site.