Refactor API calls & add state-wide login message
This commit changes the API calls to follow the existing redux
flow, rather than being limited to component state. It also adds
a state-wide login message that displays on the custom-sign-in
component to reflect status of magic link login process. This was
added specifically so that counselors or admin users who have an
admin account but not a user profile know why their login attempts
fail.