Fix static assets on staging (#484)
* Add static asset build to cibuild

* Adjust context dir for django container

* Clean up unnecessary code

* Adjust how/when to build static assets

Local setup was previously dependent on there not being a /taui/assets/ dir, that way docker would ignore the command to copy on initial build in scripts/update when the assets had not been bundled yet. However, this isn't a very reliable setup since scripts/update could be called from a local environment and /taui/assets/ has potential to exist but be empty, which would throw an error on build. Instead, this waits to build the django container until after taui is built and bundled, so regardless of dev environment the build can succeed.

* Move yarn build outside of Dockerfile