Fix profile reset on address edit during user sign up
Geocoding was updating some of the EditProfile props, triggering
componentWillReceiveProps and setting profile state using the app
state's version of the profile, removing anything the user had
changed up to that point. Initially componentWillReceiveProps was
necessary because it might be handed a profile that was null; however,
our new login process means profile should never be empty, making
componentWillReceiveProps obsolete and even problematic in cases
like this.