Make deep copy of profile for restoration on cancel
Due to the way destinations are edited and added, clicking cancel
wasn't restoring changed destinations to what they had been before
editing. This commit makes a deep copy of the user’s profile on component
render that is used to restore the entire profile, including destinations, if the
user clicks cancel.