Drop NAT Gateway
We can just use one main subnet and only assign public elastic IPs when we need to access the internet. We do need
to have a public IP on the ECS task so it can reach rollbar and ECR (could fix ECR to use PrivateLink), but we don't allow
direct ingress to it.

RDS has a specific flag to prevent it from being accessible (default to no public IP, but we will set explicitly).
https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_VPC.WorkingWithRDSInstanceinaVPC.html#USER_VPC.Hiding

Some changes are from terraform fmt.