Cap max travel time value in UI
The network analysis has a max time travel value set when generating path and times data. Because of the way Taui parses times data, a time value that exceeds the analysis max time travel gets translated as a 32 bit Int's max value: 2147483647. This does not affect the travel time scoring weight but does show up in our UI. Updates to catch the invalid max value and display 120+ in its place